+++ KkuMulKum/Source/SetReadyInfo/ViewModel/SetReadyInfoViewModel.swift
